Fix typography in Problem 2.7 (New York Point)
Two issues here:

In the solution to "Iggie", the final _e_ was quite a bit lower than it
ought to be, due to the fact that the second row of the `smallmatrix`
was completely empty. To fix this, I added a `\,` thinspace in the same
way as the other _e_ instances.

A second, more subtle issue was the fact that all instances of _n_ were
floating slightly above where they should be. In this case, adding a
thinspace in the first row was not sufficient, so instead I added a
`\vphantom{\bullet}`, and this brought everything back into alignment.